requirements.txt
openai~=1.55.2
datetime~=5.5
fastapi~=0.115.5
uvicorn~=0.32.1
requests~=2.32.3
python-dotenv~=1.0.1